pyproject.toml
[tool.poetry]
name = "mytk"
version = "0.1.0"
description = ""
authors = ["Kenta Fukuhara <kakikaku.fukuhara@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.6.1, <=3.8"
Pillow = "8.4.0"
numpy = "1.19.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
